### 2020-05-27 #2227 @humanfactors
Thank you for message @arfon, all delays are very much understood during these tough times. Quick question, are we able to start using the assigned DOI in the article proof as of now, or should we continue referencing the DOI assigned from PsyArXiv preprint?

Note, @lukestrickland and @ballardtj the article proof processed above.

## Suggested Reviewers

I've identified three potential reviewers (see below) from the list of volunteer reviewers based on expertise in both R and Cognitive Modelling/Psychology. Unfortunately, I don't know of other suitable reviewers, though am happy to expand suggestions from list as required.

- craddm
- wjakethompson
- mmrabe

Thanks for your assistance!
