### 2020-05-27 #2227 @arfon
> Thank you for message @arfon, all delays are very much understood during these tough times. Quick question, are we able to start using the assigned DOI in the article proof as of now, or should we continue referencing the DOI assigned from PsyArXiv preprint?

Please don't use the DOI from the proof - it's not the final DOI.
