### 2020-05-28 243a6c8
Add sleep.id column to vector generated dataframe
- Previously the sleep.id column was missing from a vector sequence generated FIPS_df.
- This preveented the plotting function from overlaying sleep rects, resulting in an error.
- This should fix the issue now.
