+++ R/data_singlevectorformat.R
@@ -2,6 +2,10 @@ wakesleep_seq_is_binary <- function(x) {
     return(all(x == 0 | x == 1))
 }
 
+wakesleep_seq_is_logical <- function(x) {
+    return(all(x == T | x == F))
+}
+
 validate_wakesleep_seq <- function(seq) {
     if(!wakesleep_seq_is_binary(seq)) {
         stop("Binary sequence must only contain 1's (wake) and 0's (sleep)")
@@ -51,13 +55,14 @@ parse_sleepwake_sequence <- function(seq, epoch, series.start) {
 
     # Status based sequence
     wake_status_int = seq
+    sleep.id = bitvec_to_cumcount(seq)
     wake_status = as.logical(seq)
     change_point = change_points(seq)
     switch_direction = status_dir(seq, change_point)
     status_duration = time_in_status(wake_status, epoch)
     total_prev = shifted_time_status(wake_status, change_point, epoch)
 
-    return(as_FIPS_df(data.frame(datetime, time, day, sim_hours, wake_status_int,
+    return(as_FIPS_df(data.frame(datetime,sleep.id, time, day, sim_hours, wake_status_int,
     wake_status, change_point, switch_direction, status_duration, total_prev, stringsAsFactors = F)))
 
 }
