### 2020-06-15 #2227 @humanfactors
Hi @oliviaguest, thanks for efforts reaching out to potential reviewers. I might have some ideas, the issue of course is finding individuals with overlapping subject matter familiarity/expertise as well as R expertise. I assume we need both?

Depending on flexibility with expertise areas, I might have some more suggestions. In any case, I'll conduct a broader search for suggestions.

Cheers.
