### 2020-06-15 #2340 @whedon
[REVIEW]: FIPS: An R Package for Biomathematical Modelling of Human Fatigue Related Impairment
**Submitting author:** @humanfactors (<a href="http://orcid.org/0000-0003-4143-7308">Michael David Wilson</a>)
**Repository:** <a href="https://github.com/humanfactors/FIPS" target ="_blank">https://github.com/humanfactors/FIPS</a>
**Version:** v0.1.1
