### 2020-06-18 e719c7d
Update for compatibility with dplyr v1.0.0
Attribute loss issues are resolved in dplyr 1.0.0
