+++ R/FIPS_simulation.R
@@ -86,7 +86,7 @@ FIPS_Simulation_lost_attributes <- function(x) {
 print.FIPS_simulation <- function(x) {
 
   if(FIPS_Simulation_lost_attributes(x)) {
-    warning("Your FIPS_Simulation object has lost attributes (have you wrangled the dataframe with dplyr?). Dispatching method onwards.")
+    warning("Your FIPS_Simulation object has lost attributes (have you wrangled the dataframe with dplyr (version < 1.0.0?). Dispatching method onwards.")
     NextMethod()
   } else {
     help_function = switch(get_FIPS_modeltype(x),
