### 2020-07-07 #2340 @humanfactors
Hi @oliviaguest, I assume you mean in terms of package?

Currently, we are not working on any features beyond bug fixes, and will likely remain like this for a couple of months. There are future development plans to extend the modelling interface (potentially), improve theming of the plots, and incorporate an additional model. However, this is likely to be based on demand, resources, and package reception.

More broadly, we currently have a couple of papers in progress that directly depend on FIPS, so we are regularly working with FIPS.

Not sure if that answers the question though. Thanks again for assistance thus far.
