### 2020-07-07 #2340 @oliviaguest
Ah, cool. I was wondering for this JOSS paper specifically, but that all sounds very organised — nice. Have you all @craddm, @anibalsolon, @humanfactors discussed what needs to be done for this to have all boxes ticked/checked above? ☺️
