### 2020-07-07 #2340 @craddm
Hi @humanfactors,

This seems like an interesting package, and everything I've tried with it so far has worked great. A couple of comments, in no particular order:

- I initially thought there were a few functions that don't currently have examples (e.g. `expand_sleep_series()`) that would benefit from them. But then I realized that those functions aren't exported, user-facing functions. You might want to make them internal functions (e.g. using `@keywords internal`) so that they don't show up in the user-facing API docs

- I'm not generally familiar with the area so have no example data of my own to test the package out. The package has a great vignette that generates data of the appropriate type, but one thing I'm left a little curious about is how you *normally* get data of that type. Specifically, there's a function mentioned in the pdf, `parse_sleepwake_sequence()`, that's intended for dealing with a common format used by other software for sleep/wake sequences. Examples of this type of data and the use of the function would be appreciated.

-  I can see you have some tests set up using `testthat`, which is great. But I'd also suggest hooking them up to a continuous integration service such as Travis-CI.

Matt
