### 2020-07-08 f450fad
Travis integration
First attempt at trying to get automated testing via Travis.
