### 2020-07-08 1b33013
@keywords internal flag on support functions
Add @keywords internal to parse_sleep_time support functions
