### 2020-07-08 d581967
Update walkthrough vignette with bitvector
Vignette now has extensive example about use of bitvector (or sleepwake sequence) support.
