### 2020-07-08 #2340 @humanfactors
Hi @craddm

Thanks for your feedback, very much appreciated. I've now opened a pull request (see above) with several changes to the repository to address your review comments. Feel free to add any detailed comments there.

In reference to your second point, I would like to clarify:

1. Were you suggesting any changes in the manuscript are required to make clearer typical data sources? Currently, I haven't made changes to the paper.
2. Have the changes in vignettes and documentation addressed the concerns about data sources? There are quite a number of _exact_  data file types [that come from sleep tracking tech], but in our experience almost all use the bitvector as the core outputted data.

Also cheers @oliviaguest for following up on this.
