### 2020-07-08 8e775e9
Create CONTRIBUTING.md
