### 2020-07-08 #2340 @anibalsolon
Hi all,

Thank you for the follow-up, @oliviaguest!

----

Thank you for the contribution @humanfactors I am sure several will benefit from an open-source solution of these models!
Here are my two cents:

1. Following @craddm comments, CI integration is a great addition, and I am glad you got that working already. As a suggestion, adding code coverage to the CI pipeline provides you with some information about how much of your code has been tested, giving you a perspective on what is missing. Some references for that: https://github.com/r-lib/covr

 Together with coverage, adding a linting tool helps to see what the tired eyes might not get, such as styling and misreference to/unused variables. Here is an example of how to integrate with a CI tool: https://cran.r-project.org/web/packages/lintr/readme/README.html#continuous-integration

 And here is an example of `lintr` for `R/FIPS_df.R` on RStudio:
![image](https://user-images.githubusercontent.com/562525/86938497-1ab11980-c106-11ea-8a51-fd77a0265d7d.png)

2. I saw you added the `CONTRIBUTING.md` file to the repo, which is good! It gives newcomers a better sense of how things work for your project (personally, it reduces a lot of anxiety). I would like to suggest to take a look at the Contributing guidelines from fMRIPrep, they did a great job here: https://github.com/poldracklab/fmriprep/blob/master/CONTRIBUTING.md

 I do not believe it needs to be this thorough, so I point to two main aspects of it: **Issues and Pull requests**.
  * For **issues**, I would suggest you could give some pointers on which information is needed so the collaborators can understand the problem and help the user debug it, hopefully reaching a fix. So, let's say, it would prevent one to open an issue like "It does not work!!1!1!", although people have been much more reasonable these days. Along with that, Github provides issue templates, in which you can provide some basic text for which information would smooth out the debugging process e.g. R version, Operational System, FIPS version. This is how it works: https://docs.github.com/en/github/building-a-strong-community/about-issue-and-pull-request-templates
  * For **pull requests**, you can also provide a template for the relevant information for the PR e.g. what it solves, why is it needed, etc. But also, it would be nice including to the Contributing guidelines some technical aspects like code formatting, function documentation with docstring, and how to run tests locally during development.

The manuscript is well-written, and it seems to me that it covers the required details of this package.
