### 2020-07-09 2c15590
Expand contributing guidelines
