### 2020-07-09 26aa22e
Fix non-trival violations of code formatting policy
