### 2020-07-09 c1867a4
Update CONTRIBUTING.md
