+++ CONTRIBUTING.md
@@ -43,4 +43,7 @@ A brief list of coding conventions, which are somewhat atypical:
 - Underscores (i.e., snake-case) should separate words in object names. Period separation is currently used for legacy supported functions.
 - When naming objects, abbreviations and parameter names should use capitals as a first priority (e.g., `TPM_make_pvec` not `tpm_make_pvec`; but `unified_make_pvec` is fine).
 
+Development should be conducted on a separate branch. Tests are run using the `testthat` package automatically with Rstudio default.
+This can be envoked by pressing `ctrl + shift + T` in Windows or Linux. `Roxygen2` is used for all documentation.
+
 Thankyou!
