### 2020-07-09 #2340 @humanfactors
Thank you @anibalsolon for your insightful review. I've added our changes directly to the already open pull request on the `joss-craddm` branch https://github.com/humanfactors/FIPS/pull/3

> As a suggestion, adding code coverage to the CI pipeline provides you with some information about how much of your code has been tested, giving you a perspective on what is missing. Some references for that: https://github.com/r-lib/covr

We really appreciated this suggestion, as I had often wondered how coverage reports were generated. I've now added this to the Travis CI script and also run the function locally. I'm satisfied with the coverage, and in fact was surprised it was that high. Moreover, adding this detected a missing `tests/testthat.R` file!

![image](https://user-images.githubusercontent.com/16096044/87013520-5975cb00-c1fd-11ea-8e26-d82d769d086e.png)

> Together with coverage, adding a linting tool helps to see what the tired eyes might not get, such as styling and misreference to/unused variables.

We carefully considered including a linting tool in the automated builds. Indeed, @lukestrickland and I both went through the linting suggestions run by `lintr`. We found only a small subset of the enormous number of suggestions were of significant importance, but have corrected issues that clearly violated our code style guide. In line with this, we do now recognize the need for our conventions to be mapped out more explicitly to better support potential contributors. In response to your second point, we have now added additional documentation to the `CONTRIBUTING.md` file (e.g., 120 column max).

We hope the package will experience more community growth, and will certainly revisit the need for automated linting and a custom `.lintr` specification at that future time.

> I saw you added the CONTRIBUTING.md file to the repo, which is good! It gives newcomers a better sense of how things work for your project (personally, it reduces a lot of anxiety). I would like to suggest to take a look at the Contributing guidelines from fMRIPrep, they did a great job here: https://github.com/poldracklab/fmriprep/blob/master/CONTRIBUTING.md

We have now clarified and expanded our `CONTRIBUTING.md` file, and adopted your suggestion of providing a Github Issue Template for bug reports. Once again, I appreciate your keen eye of the features of Github as I didn't know this existed and indeed may save effort following up "lower quality" bug reports.

> For pull requests, you can also provide a template for the relevant information for the PR e.g. what it solves, why is it needed, etc. But also, it would be nice including to the Contributing guidelines some technical aspects like code formatting, function documentation with docstring, and how to run tests locally during development.

The `CONTRIBUTING.md` now includes descriptions of:

- Package-specific code style
- The automated testing software run for local development `testthat`
- The automatic documentation generator (`roxygen2`)

In the spirit of transparency, we've also added a note to specify that we would prefer potential contributors to reach out to us prior to making a large pull request. This is for three reasons:

1. We do have some development efforts planned for the future. Ideally, we want to prevent any duplicated efforts.
2. Knowledge that others are interested in working on the package would help us determine how best to position a more expansive `CONTRIBUTING.md`, similar to that of `fmriprep`.
3. [Most importantly], it would make us very happy to hear other users of the package.

---------------

If there's no further additions, I will merge the open pull request into master.

Cheers,

Michael
