### 2020-07-09 24225f7
Minor paper edits (2 words)
