### 2020-07-09 #2340 @anibalsolon
That's a fine work, @humanfactors, I am really happy with its results. Thank you!

The coverage is pretty high indeed and what is not covered does not seem to impact the core features of that package. So, 🎉

I understand your point about linting, they can be nit-picky. Your details in the Contributing doc is good enough for new collaborators.

That said, the new modifications are good to go, in my opinion. Congratulations on the project, and thank you for open-sourcing it!
