### 2020-07-10 9438c21
Update paper.md
Add funding to DSTG
