### 2020-07-13 #2340 @humanfactors
Hi @oliviaguest. I've made a few minor tweaks on the master branch. I'll try build from that below.

I will just await my co-authors final check of this (should be tomorrow morning). Thank you!
