+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: FIPS
 Type: Package
 Title: Fatigue Impairment Prediction Suite (FIPS)
-Version: 0.1.0
+Version: 0.1.1
 Date: 2020-04-20
 Authors@R: c(
     person("Michael David", "Wilson", email = "michael.d.wilson@curtin.edu.au", role = c("aut", "cre")),
