### 2020-07-15 #2340 @labarba
The version tag in the repository does not reflect here. Can you confirm it should be updated to v0.1.1?
