### 2020-07-15 #2340 @humanfactors
Hi @labarba , I'll admit confusion here on my end, apologies for that.

> The version tag in the repository does not reflect here. Can you confirm it should be updated to v0.1.1?

I updated the version number in light of the changes made through the review process, so `0.1.1` is correct. I thought I should delineate between pre-review and post-review repository. Let me know if that wasn't correct step.

> Could you update the title of the Zenodo deposit to match the title of the JOSS paper? Thx.

I've changed now, I hadn't used Zenodo prior, and linked with this the release.

Cheers!
