### 2020-07-15 #2340 @labarba
I found just one typo in the caption of the figure: "indicting" for "indicating."
